% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{Expand - expand a uri template with data}
\usage{
expand(template, variables)
}
\arguments{
\item{template}{x}

\item{variables}{x}
}
\description{
Expand - expand a uri template with data
}
