% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UriTemplate.R
\name{template}
\alias{template}
\title{Returns the set of keywords in a uri template}
\usage{
template(x = "")
}
\arguments{
\item{uri}{(character) any URI}
}
\description{
Returns the set of keywords in a uri template
}
\examples{
# no URI given
x <- template()
x
x$port_mapping
x$ip_based_schemes

# give a URI
x <- template("http:.//example.com/foo/bar")
x
x$uri
## parse it
x$parse()
x$uri_parsed

# join
base <- "http://example.com"
uri <- template("relative/path")
uri$parse()
uri$join(base, uri)
}
