% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulrtemplate-package.R
\docType{package}
\name{urltemplate-package}
\alias{urltemplate}
\alias{urltemplate-package}
\title{URL Templating Following RFC6570}
\description{
URL Templating Following RFC6570
}
\author{
Scott Chamberlain
}
\keyword{package}

