% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variables}
\alias{variables}
\title{Returns the set of keywords in a uri template}
\usage{
variables(x)
}
\arguments{
\item{x}{(character) A template}
}
\value{
a character vector of keywords
}
\description{
Returns the set of keywords in a uri template
}
\examples{
variables("http:www{.domain*}{/top,next}{?q:20}")
variables("http://www.{domain}/")
variables("find{?year*}")
variables("www{.dom*}")
variables("{x,hello,y}")
variables("{#path:6}/here")
variables("up{+path}{var}/here")
}

